move loader in good folder
