+++ client/src/views/ToDoList/ToDoListDetailView.vue
@@ -8,7 +8,7 @@ import { setTitle, setDescription } from "@/utils/documentInfos.js";
 import ToDoItemFormComponent from '@/components/ToDoList/ToDoItemFormComponent.vue';
 import ToDoItemImageModalComponent from '@/components/ToDoList/ToDoItemImageModalComponent.vue';
 import ToggleComponent from '@/components/ToggleComponent.vue';
-import LoaderComponent from '@/components/LoaderComponent.vue';
+import LoaderComponent from '@/components/Loader/LoaderComponent.vue';
 import QRCodeModal from '@/components/ToDoList/ToDoListQRCodeModal.vue';
 import logger from '@/utils/logger.js';
